Added Dry Run Flag

+++ src/obsidian2anki/services/clear_service.py
@@ -4,10 +4,8 @@ import logging
 from obsidian2anki.core.note_processor import NoteProcessor
 from obsidian2anki.core.state_manager import StateManager
 from obsidian2anki.core.vault_manager import VaultManager
-from obsidian2anki.config import Settings, get_settings
 
 logger = logging.getLogger(__name__)
-settings: Settings = get_settings()
 
 
 class ClearService:
@@ -18,10 +16,14 @@ class ClearService:
         self._state = state
         self._note_processor = note_processor
 
-    def clear(self) -> None:
+    def clear(self, dry_run: bool = False) -> None:
         state_items = self._state.get_state_items()
 
         logger.info("Founded %d notes in state", len(state_items))
+
+        if dry_run:
+            logger.debug("Ended clearing on dry run.")
+            return
 
         for id, info in state_items:
             self._note_processor.delete_note_cards(id, info.anki_note_ids)
